requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-openai
langchain-community
openai
rich
python-dotenv
jinja2